Added Doubleshot.lock. Simplified bootstrapping a bit.
Now bootstrapping generates a .classpath.cache, which is the same file generated by the normal setup process, so all that
code can be reused.

Since bootstrapping or not, we now have a Lockfile, and the build command runs Doubleshot#setup first which populates JarDependency#path for the JARs in our lockfile, the build command has been modified to use to use those paths instead of the previously un-specced Doubleshot#classpath method.